refactor: move path to top level
